Fetching keys from log lines
Keys can be fetched as a set, similar to `jsonb_object_keys`, or as an
array. Keys are returned in the order the appear, reading left-to-right.

Our functions are expected to return the same result for a given log
line. Marking them immutable allows them to be used in indexes.